refactor: 사이드바 렌더링 구현함수 수정
