+++ src/components/sideBar/index.ts
@@ -1,4 +1,5 @@
-import { addSideBarEvent, renderSideBar } from "./sideBarView";
+import { renderSideBar } from "./sideBarController";
+import { addSideBarEvent } from "./sideBarView";
 
 export const initSideBar = () => {
   renderSideBar() //
